Collect the ReadJson logic.
